requirements.txt
aescipher==6.0.0
alabaster==0.7.16
astroid==3.1.0
asttokens==2.4.1
Babel==2.14.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
DateTime==5.5
decorator==5.1.1
dill==0.3.8
docutils==0.20.1
easyrsa==3.1.0
encryptedsocket==3.0.0
executing==2.0.1
flake8==7.0.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
idna==3.6
imagesize==1.4.1
iniconfig==2.0.0
ipython==8.22.2
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
msal==1.28.0
munch==4.0.0
numpy==1.26.4
omnitools==0.0.268
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.5.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pyflakes==3.2.0
pyftpdlib==1.5.9
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.1.0
pyOpenSSL==24.1.0
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
selenium==3.141.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
threadwrapper==0.14.0
tomlkit==0.12.4
traitlets==5.14.2
types-requests==2.31.0.20240406
types-waitress==2.1.4.20240106
unencryptedsocket==2.2.0
urllib3==2.2.1
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
zope.interface==6.3
